Separated the HTTP client configuration into its own class
Also, the two newly introduced interceptors have been added to the
HTTP client interceptor chain
