+++ src/main/java/usi/si/seart/config/MainConfig.java
@@ -1,6 +1,5 @@
 package usi.si.seart.config;
 
-import okhttp3.OkHttpClient;
 import org.jetbrains.annotations.NotNull;
 import org.springframework.boot.web.servlet.FilterRegistrationBean;
 import org.springframework.context.annotation.Bean;
@@ -15,7 +14,6 @@ import usi.si.seart.converter.SupportedLanguageToDtoConverter;
 
 import java.text.DateFormat;
 import java.text.SimpleDateFormat;
-import java.util.concurrent.TimeUnit;
 
 @Configuration
 public class MainConfig {
@@ -23,15 +21,6 @@ public class MainConfig {
     @Bean
     public DateFormat utcTimestampFormat() {
         return new SimpleDateFormat("yyyy-MM-dd'T'hh:mm:ss");
-    }
-
-    @Bean
-    public OkHttpClient httpClient() {
-        return new OkHttpClient.Builder()
-                .connectTimeout(1, TimeUnit.MINUTES)
-                .writeTimeout(1, TimeUnit.MINUTES)
-                .readTimeout(1, TimeUnit.MINUTES)
-                .build();
     }
 
     @Bean
